+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Run unit tests on Windows PowerShell.
 - Run integration tests.
+- `Install-ModulePatch`
+  - Tested using integration test.
 
 ### Fixed